Display message when the payment fails

+++ local/modules/MultipleFullfilmentCenters/Listener/OrderLocalPickup.php
@@ -48,7 +48,8 @@ class OrderLocalPickup extends BaseAction implements EventSubscriberInterface
 					->filterByCartId($order->getCartId())
 					->findOne();
 				
-				$cartProductLocation->delete(); 
+				if($cartProductLocation)
+					$cartProductLocation->delete(); 
 			}
 		}
